lmsasm: add quirks for EV3-G compatibility
The bytecodes compiled by the EV3-G desktop software have a  few differences from the compiler in lms2012. In EV3-G, float constant values are always 5 bytes and the size of label offests is optimized to the smallest size.

This adds two flags for the two quirks and adds a command line argument to enable them.