assembler: add type checking
This adds basic type checking of opcode arguments.

It also adds a new compiler option for allowing unsigned integer literals - needed for existing tests and it implements the bang (!) operator to allow ignoring type checking.

The bang (!) operator is added to the .lms test programs to fix compiler errors. There are some obvious bugs in these programs that should be fixed.

On behalf of LEGO System A/S